feat(atlas-ui): moving bbs and pts files
