+++ packages/theming/atlas/src/themesource/atlas_nativemobile_content/native/buildingblocks/card.ts
@@ -1,5 +1,5 @@
 import { Platform } from "react-native";
-import { background, border, contrast, spacing } from "../variables";
+import { background, border, contrast, spacing } from "../../atlas_core/native/variables";
 /*
 ==========================================================================
     Cards
