+++ packages/theming/atlas/src/themesource/atlas_nativemobile_content/native/buildingblocks/header.ts
@@ -1,4 +1,4 @@
-import { background, spacing } from "../variables";
+import { background, spacing } from "../../atlas_core/native/variables";
 /*
 ==========================================================================
     Cards
