+++ packages/theming/atlas/src/themesource/atlas_nativemobile_content/native/pagetemplates/inspectiondetails.ts
