+++ packages/theming/atlas/src/themesource/atlas_nativemobile_content/native/pagetemplates/maps.ts
@@ -1,4 +1,4 @@
-import { background, border, contrast, spacing } from "../variables";
+import { background, border, contrast, spacing } from "../../atlas_core/native/variables";
 /*
 ==========================================================================
     Maps
