+++ packages/theming/atlas/src/themesource/atlas_web_content/web/buildingblocks/_alert.scss
